build: doc: reduce doxygen.cfg.in to non-default entries only
Also add EXCLUDE_SYMBOLS for structs that are undocumented (otherwise
doxygen makes html pages for them but warns they and their members are not
documented).

This formerly 5KB file is reduced almost down to 1/8th original size.

In the interests of a clean diff, the new file still has lines with trailing
spaces.

Signed-off-by: Duncan Roe <duncan_roe@optusnet.com.au>
Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>